% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{compare_calls}
\alias{compare_calls}
\title{Compara Duas Chamadas}
\usage{
compare_calls(call1, call2, itens)
}
\arguments{
\item{call1, call2}{chamadas a comparar}

\item{itens}{vetor de strings indicando os argumentos das calls a comparar}
}
\value{
Apenas executa os testes, sem retornar valor algum
}
\description{
Funcao interna para uso nos testes de update, nao deve ser utilizada diretamente
}
