% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{cov2}
\alias{cov2}
\title{Calculo da covariancia normalizada por 1/n}
\usage{
cov2(...)
}
\arguments{
\item{mat}{\link{matriz numerica} serie cuja covariancia se deseja calcular}
}
\value{
\link{escalar numerico} covariancia calculada
}
\description{
Funcao alternativa para calcular covariancia atraves do estimador "n"
}
