% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod-ss_reg_din.r
\name{CV_regdin}
\alias{CV_regdin}
\title{Validacao Cruzada De Regressao Dinamica}
\usage{
CV_regdin(
  serie,
  regdata,
  formula,
  vardin = FALSE,
  estatica = FALSE,
  cv_control = list()
)
}
\arguments{
\item{serie}{série para ajustar}

\item{regdata}{\code{data.frame}-like contendo variáveis explicativas}

\item{formula}{opcional, fórmula da regressão. Se for omitido, todas as variaveis em
\code{regdata} serão utilizadas}

\item{vardin}{booleano ou inteiro indicando se deve ser estimado modelo com heterocedasticidade.
Caso \code{TRUE} tenta pegar a sazonalidade da série; se for um número inteiro assume este
valor como a sazonalidade}

\item{estatica}{booleano indicando se a regressao deve ser de coeficientes estaticos}

\item{cv_control}{opcional, uma lista de controle da validacao cruzada. Ver Detalhes}
}
\value{
valor otimo de \code{lambda}
}
\description{
Realiza validacao cruzada de regressao dinamica por grid check
}
\details{
Esta funcao executa uma regressao dinamica em janela movel com diferentes valores de penalidade
\code{lambda} buscando aquele que leva ao melhor modelo do ponto de vista de previsao.

A lista \code{cv_control} pode conter um ou mais dos seguintes elementos nomeados

\describe{
\item{janela}{tamanho da janela para ajuste -- similar ao argumento homonimo em
\code{\link{janelamovel}}}
\item{passo}{salto entre uma janela e outra -- similar ao argumento homonimo em
\code{\link{janelamovel}}}
\item{refit.cada}{inteiro indicando de quantas em quantas janelas o modelo deve ser reajustado --
similar ao argumento homonimo em \code{\link{janelamovel}}}
\item{obj.ahead}{vetor de inteiros indicando a janela de passos a frente de interesse para a
previsao}
\item{error_func}{funcao que recebe um vetor de erros de previsao e retorna um escalar como
metrica de performance. Default function(x) sum(x^2)}
\item{lambda}{vetor de lambdas a serem testados}
}
}
