% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{datregdin}
\alias{datregdin}
\title{Dado Exemplo De Velocidade Do Vento}
\format{
\code{data.frame} de duas colunas:
\describe{
\item{obs}{velocidade de vento observada}
\item{prev}{velocidade de vento prevista}
}
}
\usage{
datregdin
}
\description{
Dado para exemplo de ajuste dos modelos, contendo informacoes de velocidade do vento prevista e
observada
}
\keyword{datasets}
