% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod-dcs_reg_din.r
\name{default_init_dcs}
\alias{default_init_dcs}
\title{Inicializacao Padrao De \code{dcs_reg_din}}
\usage{
default_init_dcs(coefs, serie, regdata, formula, vardin, num.obs = 0.2)
}
\arguments{
\item{coefs}{coeficientes do modelo como retornado por \code{\link[DCS]{coef.DCSmodel}}}

\item{serie}{série para ajustar}

\item{regdata}{\code{data.frame}-like contendo variáveis explicativas}

\item{formula}{opcional, fórmula da regressão. Se for omitido, todas as variaveis em
\code{regdata} serão utilizadas}

\item{vardin}{booleano indicando se deve ser estimado modelo com heterocedasticidade.}

\item{num.obs}{numero de observacoes iniciais para usar na heuristica. Se for menor do que 1, se
assume que e um percentual do tamanho da serie}
}
\value{
lista de dois elementos: valores iniciais para otimizacao e vetor de hiperparametros
fixos
}
\description{
Funcao default para inicializacao dos hiperparametros em modelos \code{dcs_reg_din}
}
\details{
A funcao de inicializacao padrao implementa uma heuristica relativamente simples para
inicializacao dos hiperparametros. A partir das \code{num.obs} primeiras observacoes da serie e
variaveis explicativas, estima uma regressao linear simples com a mesma especificacao da que sera
modelada dinamicamente e usa estes coeficientes como valores iniciais para a otimizacao dos
hiperparametros em \code{dcs_reg_din}. Estes valores sao entao removidos da otimizacao, isto e,
sao tratados como fixos (geralmente nao e um problema, ainda mais com series mais longas).

Se \code{vardin = TRUE}, tambem coloca os scores dos ciclos de variancia como valores fixos em 0,
gerando assim heterocedasticidade deterministica.
}
