% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classe-modprevP.r
\name{estimamodelo_P}
\alias{estimamodelo_P}
\title{Estimacao De Modelos Periodicos}
\usage{
estimamodelo_P(serie, tipo, ...)
}
\arguments{
\item{serie}{série para ajustar}

\item{tipo}{tipo de modelo a ser ajustado. Ver Detalhes}

\item{...}{demais parâmetros passados para as funções de fit específicas de cada modelo}
}
\value{
Objeto da classe modprev e subclasse igual a \code{tipo}, uma lista de dois elementos:
\code{modelo} e \code{serie} contendo o modelo estimado e a série passada
}
\description{
Wrapper para estimacao de modelos periodicos
}
\details{
Esta funcao nao deve ser chamada diretamente pelo usuario, mas sim internamente por
\code{\link{estimamodelos}} quando o argumento \code{periodico == TRUE}. Descricoes mais
detalhadas da estimacao de modelos, periodicos ou nao, devem ser buscadas em
\code{\link{estimamodelos}} diretamente.

Modelos periodicos sao estimados e utilizados em cima do arcabouço ja desenvolvido para os
modelos unicos. Esta classe nada mais e do que um wrapper em torno de uma lista de modelos, um
para cada estacao da serie, estimados por \code{\link{estimamodelo_U}}. Todos os metodos sao
desenvolvidos em torno daqueles ja existentes para cada arcabouco de modeloagem, sendo a classe
\code{modprevP} um simples coordenador do uso destes S modelos periodicos.
}
