% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classe-modprevS.r
\name{estimamodelo_S}
\alias{estimamodelo_S}
\title{Estimacao De Modelos Com Shapesifting}
\usage{
estimamodelo_S(target_pipe, data_pipe, data_list, tipo, ...)
}
\arguments{
\item{target_pipe}{pipes brutos indicando como extrair a serie alvo a ser modelada}

\item{data_pipe}{pipes brutos indicando como extrair as variaveis explicativas a serem usadas}

\item{data_list}{lista nomeada contendo os dados a serem usados na extracao via pipes}

\item{tipo}{tipo de modelo a ser ajustado. Ver \code{\link{estimamodelo}}}

\item{...}{demais parâmetros passados para as funções de fit específicas de cada modelo}
}
\value{
Objeto da classe \code{modprev_S}
}
\description{
Wrapper para estimacao de modelos com shapeshifting
}
\details{
Esta funcao nao deve ser chamada diretamente pelo usuario, mas sim internamente por
\code{\link{estimamodelo}} quando os argumentos \code{data_pipe}, \code{target_pipe} e \code{data_list} sao usados.
Descricoes mais detalhadas da estimacao de modelos, periodicos ou nao, devem ser buscadas em
\code{\link{estimamodelo}} diretamente.
}
