% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classe-modprevU.r
\name{estimamodelo_U}
\alias{estimamodelo_U}
\title{Estimacao De Modelos Unicos}
\usage{
estimamodelo_U(serie, tipo, ...)
}
\arguments{
\item{serie}{serie para ajustar}

\item{tipo}{tipo de modelo a ser ajustado. Ver Detalhes}

\item{...}{demais parâmetros passados para as funções de fit específicas de cada modelo. Ver
Detalhes}
}
\value{
Objeto da classe modprevU e subclasse igual a \code{tipo}, uma lista de dois elementos:
\code{modelo} e \code{serie} contendo o modelo estimado e a série passada
}
\description{
Wrapper para estimacao de modelos unicos
}
\details{
Esta funcao nao deve ser chamada diretamente pelo usuario, mas sim internamente por
\code{\link{estimamodelos}} quando o argumento \code{periodico == FALSE}. Descricoes mais
detalhadas da estimacao de modelos, periodicos ou nao, devem ser buscadas em
\code{\link{estimamodelos}} diretamente.
}
\examples{

}
\seealso{
Other Metodos modprevU: 
\code{\link{predict.modprevU}()},
\code{\link{update.modprevU}()}
}
\concept{Metodos modprevU}
