% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/janelamovel.r
\name{expanderefit}
\alias{expanderefit}
\title{Indices De Tempo Para Reajustar Modelo}
\usage{
expanderefit(janelas, refit.cada)
}
\arguments{
\item{janelas}{lista de inicios e finais de janelas retornada por \code{\link{expandejanelas}}}

\item{refit.cada}{inteiro indicando de quantos em quantos instantes o modelo deve ser reajustado}
}
\value{
vetor logico do mesmo tamanho de \code{janelas} indicando em quais janelas o modelo deve
ser reestimado
}
\description{
Interpreta os argumentos passados a \code{\link{janelamovel}} a respeito dos reajustes
}
\details{
O vetor retornado sempre tera a primeira posicao igual a \code{FALSE}, pois no ambito da
janelamovel a primeira janela sempre corresponde a primeira estimacao do modelo, uma que e feita
por fora da janela movel.
}
