% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.r
\name{extraiserie}
\alias{extraiserie}
\title{Extrator de Janela do Verificado}
\usage{
extraiserie(
  dat,
  data_ini,
  hora_ini = "00:00",
  data_fim,
  hora_fim = "23:30",
  num_dias
)
}
\arguments{
\item{dat}{data.frame padrao contendo coluna de data e 48 colunas de valores (uma por meia hora)}

\item{data_ini}{inteiro (formato numerico aaaammdd) indicando data inicial da janela}

\item{hora_ini}{string (formato "HH:MM") indicando hora inicial da janela. Padrao "00:00"}

\item{data_fim}{inteiro (formato numerico aaaammdd) indicando data final da janela}

\item{hora_fim}{string (formato "HH:MM") indicando hora final da janela. Padrao "23:00"}

\item{num_dias}{parametro opcional, se fornecido deve ser um inteiro indicando número total de
dias para extracao}
}
\value{
serie temporal na janela especificada
}
\description{
Função para facilitar isolamento de uma parte da série de históricos no formato padrão
}
\details{
\emph{Função deprecada, sera removida na proxima versao}

O formato padrao ao qual se refere é aquele mais tipicamente encontrado nos históricos dos
modelos de eólica e solar, isto é, data.frames de 49 colunas onde a primeira corresponde a data
e as seguintes valores por meia hora do dia\tabular{llllll}{
   V1 \tab V2 \tab V3 \tab V4 \tab ... \tab V49 \cr
   20170101 \tab XX \tab XX \tab XX \tab ... \tab XX \cr
   20170102 \tab XX \tab XX \tab XX \tab ... \tab XX \cr
   20170103 \tab XX \tab XX \tab XX \tab ... \tab XX \cr
}


Existem duas formas de especificar a janela. A primeira e mais simples consiste em passar ambas
as datas e horas limite \code{data_ini}, \code{hora_ini}, \code{data_fim}  \code{hora_fim}. Uma
segunda forma alternativa permite a especificação de apenas uma ponta da janela em combinação com
o argumento opcional \code{num_dias}, de modo que o outro limite da janela será calculado de
acordo. Caso todos os parâmetros sejam fornecidos, \code{num_dias} será ignorado
}
