% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-registry.r
\name{get_registry}
\alias{get_registry}
\title{Model Registry}
\usage{
get_registry()
}
\value{
Object of class \code{"model_registry"} with component \code{models}: a named list of
\link{model_spec} objects.
}
\description{
Access the global model registry containing all registered model specifications.
}
\examples{
get_registry()

}
\seealso{
\code{\link[=register_model]{register_model()}}, \code{\link[=get_model]{get_model()}}, \code{\link[=list_models]{list_models()}}, \link{model_spec}

Other model_registry: 
\code{\link{get_model}()},
\code{\link{is_registered}()},
\code{\link{list_models}()},
\code{\link{model_spec}()},
\code{\link{register_model}()}
}
\concept{model_registry}
