% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod-par.r
\name{idordem}
\alias{idordem}
\title{Identificação De Ordem Automática}
\usage{
idordem(serie, m, max.p, A12 = FALSE, medias = NULL)
}
\arguments{
\item{serie}{serie temporal}

\item{m}{o mes para o qual identificar a ordem}

\item{max.p}{maxima ordem para testar na identificacao}

\item{A12}{booleano indicando o uso de parcela de medias moveis}

\item{medias}{opcional, serie temporal de medias moveis. So tem uso caso \code{A12 = TRUE}}
}
\value{
escalar inteiro correspondendo à ordem identificada
}
\description{
Função interna para identificação da ordem de um modelo periódico individual
}
