% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod-lgbm.r
\name{light_gbm}
\alias{light_gbm}
\title{Modelos \code{LGBM}}
\description{
Estimação e métodos de modelos da classe \code{LGBM}
}
\details{
Boosting de árvores pelo algoritmo lgbm estimado atraves de \code{\link[lightgbm]{lgbm}}. Para
mais detalhes a respeito desta modelagem e seu uso, veja a documentacao oficial do pacote.
}
