% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{match_fun_args}
\alias{match_fun_args}
\title{Match De Argumentos Com Lista}
\usage{
match_fun_args(args, fun)
}
\arguments{
\item{args}{lista nomeada contendo candidatos a argumentos de uma funcao}

\item{fun}{funcao cujos arugmentos procurar em \code{args}}
}
\value{
\code{args} reduzida aos elementos que correspondem a argumentos de \code{fun}
}
\description{
Identifica numa lista \code{args} quais elementos sao arugmentos de uma funcao \code{fun}
}
