% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{medias_sazo}
\alias{medias_sazo}
\title{Médias Móveis Sazonais}
\usage{
medias_sazo(serie)
}
\arguments{
\item{serie}{a serie da qual calcular media movel}
}
\value{
serie temporal com mesmos atributos que \code{serie} contendo as médias
}
\description{
Calcula a médias móveis rolantes para uso nos modelos PAR(p) com parcela anual
}
\details{
As médias móveis aqui calculadas são realizadas com largura de janela igual ao período de
sazonalidade da série. As observações na série de média correspondem às médias dos \code{s} meses
anteriores à observação correspondente em \code{serie}
}
