% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod-boost.r
\name{modelos_boost}
\alias{modelos_boost}
\alias{BOOST}
\alias{predict.BOOST}
\alias{update.BOOST}
\title{Modelos \code{BOOST}}
\usage{
BOOST(
  serie,
  regdata,
  formula,
  family = "Gaussian",
  cv_control = list(tipo = "kfold", B = floor(nrow(regdata) * 0.8), nthreads = 1, async =
    TRUE),
  ...
)

\method{predict}{BOOST}(object, newdata, n.ahead, ...)

\method{update}{BOOST}(object, newseries, newregdata, refit = FALSE, ...)
}
\arguments{
\item{serie}{série para ajustar}

\item{regdata}{\code{data.frame}-like contendo variáveis explicativas}

\item{formula}{opcional, fórmula da regressão. Se for omitido, todas as variaveis em
\code{regdata} serão utilizadas}

\item{family}{string indicando uma familia para funcao perda. Veja \code{\link[mboost]{mboost}}
para mais detalhes}

\item{...}{nao possui uso, existe apenas para consistencia com a generica}

\item{object}{objeto com classes \code{c("BOOST", "modprev")} contendo modelo}

\item{newdata}{\code{data.frame}-like contendo variéveis explicativas fora da amostra}

\item{n.ahead}{número de passos à frente para previsão. Este argumento não é necessario, caso não
seja informado a previsão sera feita tantos passos à frente quanto amostras em \code{newdata}}

\item{newseries}{nova série com a qual atualizar o modelo}

\item{newregdata}{\code{data.frame}-like contendo variáveis explicativas na nova amostra}

\item{refit}{booleano indicando se o modelo deve ou nao ser reajustado}
}
\value{
Objeto da classe \code{modprev} e subclasse \code{BOOST}, uma lista de dois
elementos: \code{modelo} e \code{serie} contendo o modelo estimado e a série passada

\code{predict} serie temporal multivariada contendo valor esperado e desvio padrao de
previsão \code{n.ahead} passos à frente;

\code{update} retorna modelo com novos dados e, caso \code{refit == TRUE}, reajustado.
Contrário à função de estimação, \code{update} já retorna o objeto da classe \code{modprev};
}
\description{
Estimação e métodos de modelos da classe \code{BOOST}

Para estimação deste tipo de modelo, além de ser fornecido o arumento \code{serie} também deve
ser necessariamente informado \code{regdata}, um \code{data.frame}-like contendo as variáveis
explicativas necessárias.
}
\details{
Boosting de modelos, estimados atraves da funcao \code{\link[mboost]{mboost}}. Para mais
detalhes a respeito desta modelagem e seu uso, veja a documentacao oficial do pacote.

Opcionalmente pode ser informado o argumento \code{formula} contendo a especificação do modelo
aditivo no formato compativel (veja \code{\link[mboost]{mboost}}), porem sem o LHS. Caso
\code{formula} seja omitido, todas as variaveis serao utilizadas.

Na execucao em janela movel, o vetor de pesos deve ser passado com o tamanho igual ao da janela.
O mesmo vetor sera aplicado em todas as janelas avaliadas.
}
