% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod-dcs_reg_din.r
\name{modelos_dcs_reg_din}
\alias{modelos_dcs_reg_din}
\alias{dcs_reg_din}
\alias{predict.dcs_reg_din}
\alias{update.dcs_reg_din}
\title{Modelos \code{dcs_reg_din}}
\usage{
dcs_reg_din(
  serie,
  regdata,
  formula,
  d = 1,
  vardin = FALSE,
  init.fun = default_init_dcs,
  ...
)

\method{predict}{dcs_reg_din}(object, newdata, n.ahead, S = 5000, ...)

\method{update}{dcs_reg_din}(object, newseries, newregdata, refit = FALSE, ...)
}
\arguments{
\item{serie}{série para ajustar}

\item{regdata}{\code{data.frame}-like contendo variáveis explicativas}

\item{formula}{opcional, fórmula da regressão. Se for omitido, todas as variaveis em
\code{regdata} serão utilizadas}

\item{d}{expoente da matrix de informação de Fisher no cálculo do score padronizado}

\item{vardin}{booleano indicando se deve ser estimado modelo com heterocedasticidade.}

\item{init.fun}{uma funcao que retorne inicializacao dos hiperparametros. Ver Detalhes}

\item{...}{nao possui uso, existe apenas para consistencia com a generica}

\item{object}{objeto com classes \code{c("dcs_reg_din", "modprev")} contendo modelo}

\item{newdata}{\code{data.frame}-like contendo variéveis explicativas fora da amostra}

\item{n.ahead}{número de passos à frente para previsão. Este argumento não é necessario, caso não
seja informado a previsão sera feita tantos passos à frente quanto amostras em \code{newdata}}

\item{S}{inteiro indicando o numero de simulações a serem realizadas}

\item{newseries}{nova série com a qual atualizar o modelo}

\item{newregdata}{\code{data.frame}-like contendo variáveis explicativas na nova amostra}

\item{refit}{booleano indicando se o modelo deve ou nao ser reajustado}
}
\value{
Objeto da classe \code{modprev} e subclasse \code{dcs_reg_din}, uma lista de dois
elementos: \code{modelo} e \code{serie} contendo o modelo estimado e a série passada

\code{predict} serie temporal multivariada contendo valor esperado e desvio padrao de
previsão \code{n.ahead} passos à frente;

\code{update} retorna modelo com novos dados e, caso \code{refit == TRUE}, reajustado.
Contrário à função de estimação, \code{update} já retorna o objeto da classe \code{modprev};
}
\description{
Estimação e métodos de modelos da classe \code{dcs_reg_din}

Os modelos de regressão dinâmica estão entre os mais complexos deste pacote. Para sua estimação,
além de ser fornecido o arumento \code{serie} também deve ser necessariamente informado
\code{regdata}, um \code{data.frame}-like contendo as variáveis explicativas necessárias.

A previsão destes modelos é feita por simulação, ou seja, é representada pela média de \code{S}
cenários simulados \code{n.ahead} passos à frente
}
\details{
Opcionalmente pode ser informado o argumento \code{formula} contendo a especificação da regressão
linear, no formato padrão do R (veja \code{\link{formula}}) porém sem o LHS. Caso \code{formula}
seja omitido, todas as variáveis em \code{regdata} serão utilizadas aditivamemte, i.e. se existem
as colunas \code{c("V1", "V2", "V3")}, formula sera \code{~ V1 + V2 + V3}.

\code{vardin} controla se o modelo será estimado com variâncias variantes no tempo ou não. Caso
seja \code{FALSE} um modelo homocedástico é estimado; \code{TRUE} tenta buscar sazonalidade nos
atributos de \code{serie} para heterocedasticidade e, por fim, se \code{vardin} for um inteiro
utiliza este número como sazonalidade.

As variâncias variantes no tempo são modeladas como função de variáveis circulares determinadas
a partir da sazonalidade. Isto garante que os valores sejam consistentes entre si e não aumenta
muito o número de hiperparâmetros sendo estimados

A função \code{init.fun} deve ser da forma
\code{function(coefs, serie, regdata, formula, vardin, ...)}, em que \code{...} são demais
argumentos que ela possa precisar, passados por \code{...} em \code{dcs_reg_din}. Esta funcao
deve sempre retornar uma lista de dois elementos: o primeiro deve ser o vetor de valores iniciais
dos hiperparâmetros; o segundo um vetor indicando quais devem ser tratados como valores fixos.
Por padrão é usada a função \code{\link{default_init_dcs}}.


}
