% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod-lgbm.r
\name{predict.LGBM}
\alias{predict.LGBM}
\title{\bold{Predict}:}
\usage{
\method{predict}{LGBM}(object, newdata, n.ahead, ...)
}
\arguments{
\item{object}{objeto com classes \code{c("LGBM", "modprev")} contendo modelo}

\item{newdata}{\code{data.frame}-like contendo variéveis explicativas fora da amostra}

\item{n.ahead}{número de passos à frente para previsão. Este argumento não é necessario, caso não
seja informado a previsão sera feita tantos passos à frente quanto amostras em \code{newdata}}
}
\value{
\code{predict} serie temporal multivariada contendo valor esperado e desvio padrao de
previsão \code{n.ahead} passos à frente;
}
\description{
\bold{Predict}:
}
