% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod-par.r
\name{modelos_par}
\alias{modelos_par}
\alias{PAR}
\alias{PAR_A}
\alias{predict.PAR}
\alias{update.PAR}
\title{Modelos \code{PAR}}
\usage{
PAR(serie, ...)

PAR_A(serie, ...)

\method{predict}{PAR}(object, n.ahead, ...)

\method{update}{PAR}(object, newseries, refit = FALSE, ...)
}
\arguments{
\item{serie}{série para ajustar}

\item{...}{para \code{update} nao tem uso, existe apenas para consistência com a genérica}

\item{object}{objeto com classes \code{c("PAR"|"parA", "modprev")} contendo modelo}

\item{n.ahead}{número de passos à frente para previsão}

\item{newseries}{nova série com a qual atualizar o modelo}

\item{refit}{booleano indicando se o modelo deve ou não ser reajustado}
}
\value{
Objeto da classe \code{modprev} e subclasse \code{PAR}, uma lista de dois
elementos: \code{modelo} e \code{serie} contendo o modelo estimado e a série passada

\code{predict} retorna uma série temporal multivariada contendo valor esperado e desvio
padrão da previsão \code{n.ahead} passos à frente;

\code{update} retorna modelo com novos dados e, caso \code{refit == TRUE}, reajustado.
Contrário à função de estimação, \code{update} já retorna o objeto da classe \code{modprev};
}
\description{
Estimação e métodos de modelos da classe \code{PAR}

\code{...} pode conter qualquer, ou nenhum, argumento de \code{\link[parmodels:par]{parmodels::par()}} exceto por \code{serie}, que
sera passado automaticamente como \code{serie} desta funcao.
}
\details{
Modelos periódicos autorregressivos, possivelmente com parcela anual, similares aos do GEVAZP
}
