% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod-par.r
\name{modelos_parp}
\alias{modelos_parp}
\alias{parp}
\title{Modelos \code{parp}}
\usage{
parp(serie, s = frequency(serie), p = "auto", A12 = FALSE, max.p = 11, ...)
}
\arguments{
\item{serie}{série para ajustar}

\item{s}{periodo de sazonalidade, caso \code{serie} nao seja uma serie temporal com sazonalidade}

\item{p}{vetor de comprimento igual a sazonalidade contendo as ordens de cada modelo periódico;
caso tenha comprimento menor que \code{s}, sera reciclado. Ver Detalhes}

\item{A12}{booleano indicando se a parcela anual deve ser incorporada à estimação}

\item{max.p}{vetor de comprimento igual a sazonalidade contendo as máximas ordens para
identificação automática; será reciclado se necessário. Ignorado se \code{p = "auto"}}

\item{...}{nao possui uso, existe apenas para consistencia com a generica}
}
\value{
Objeto da classe \code{modprev} e subclasse \code{parp}, uma lista de dois
elementos: \code{modelo} e \code{serie} contendo o modelo estimado e a série passada
}
\description{
Estimação e métodos de modelos da classe \code{parp}

Modelos PAR(p) são necessariamente ajustados a séries temporais com sazonalidade. Desta forma, o
argumento \code{serie} deve ser um objeto \code{ts} com sazonalidade ou então o período de
sazonalidade deve ser informado através de \code{s}. Caso nenhuma destas condições seja atendida
a função aborta com erro.
}
\details{
Modelos periódicos autorregressivos, possivelmente com parcela anual, similares aos do GEVAZP

A identificação de ordem destes modelos pode ser feita de forma automática, informando o
argumento \code{p = "auto"}. A seleção da ordem de cada modelo é realizada a partir das PACFs
periódicas, buscando o primeiro lag, a partir do primeiro, dentro do intervalo de confiança para
não significância, até o lag \code{max.p} correspondente. Para consistência com o GEVAZP, a
identificação automática usa funções deautocorrelação parcial periódica quando \code{A12 = FALSE}
e condicional do contrário.
}
