% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod-reg_lin.r
\name{modelos_reg_lin}
\alias{modelos_reg_lin}
\alias{reg_lin}
\title{Modelos \code{reg_lin}}
\usage{
reg_lin(serie, regdata, formula, ...)
}
\arguments{
\item{serie}{série para ajustar}

\item{regdata}{\code{data.frame}-like contendo variáveis explicativas}

\item{formula}{opcional, fórmula da regressão. Se for omitido, todas as variaveis em
\code{regdata} serão utilizadas}
}
\value{
Objeto da classe \code{modprev} e subclasse \code{reg_lin}, uma lista de dois
elementos: \code{modelo} e \code{serie} contendo o modelo estimado e a série passada
}
\description{
Estimação e métodos de modelos da classe \code{reg_lin}
}
