% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod-sarima.r
\name{modelos_sarima}
\alias{modelos_sarima}
\alias{sarima}
\alias{predict.sarima}
\alias{update.sarima}
\title{Modelos \code{sarima}}
\usage{
sarima(serie, ...)

\method{predict}{sarima}(object, n.ahead, ...)

\method{update}{sarima}(object, newseries, refit = FALSE, ...)
}
\arguments{
\item{serie}{serie para ajustar}

\item{object}{objeto com classes \code{c("sarima", "mod_eol")} contendo modelo}

\item{n.ahead}{numero de passos a frente para previsao}

\item{newseries}{nova serie com a qual atualizar o modelo}

\item{refit}{booleano indicando se o modelo deve ou nao ser reajustado}
}
\value{
\code{sarima} retorna modelo sarima estimado;

\code{predict} serie temporal multivariada contendo valor esperado e desvio padrao de
previsao;

\code{update} retorna modelo com novos dados e, caso \code{refit == TRUE}, reajustado;
}
\description{
Estimacao e metodos de modelos da classe \code{sarima}
}
