% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod-sarimax.r
\name{modelos_sarimax}
\alias{modelos_sarimax}
\alias{sarimax}
\title{Modelos \code{sarimax}}
\usage{
sarimax(serie, regdata = NULL, formula = NULL)
}
\arguments{
\item{serie}{serie para ajustar}

\item{regdata}{\code{data.frame}-like contendo variáveis explicativas}

\item{formula}{formula da regressão. Se for omitido, todas as variaveis em \code{regdata} serão
utilizadas. So tem uso se \code{regdata} for passado}
}
\value{
Objeto da classe \code{modprev} e subclasse \code{sarimax}, uma lista de dois elementos:
\code{modelo} e \code{serie} contendo o modelo estimado e a série passada
}
\description{
Estimação e métodos de modelos da classe \code{sarimax}
}
\details{
Esta especificação identifica e ajusta um modelo SARIMAX(p, d, q)(P, D, Q)^s através da função
\code{\link[forecast]{auto.arima}}, que realiza uma busca no espaço de hiperparâmetros de ordem
do modelo por aquele conjunto com menor BIC.
}
