% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod-ss_ar1_saz.r
\name{modelos_ss_ar1_saz}
\alias{modelos_ss_ar1_saz}
\alias{ss_ar1_saz}
\alias{predict.ss_ar1_saz}
\alias{update.ss_ar1_saz}
\title{Modelos \code{ss_ar1_saz}}
\usage{
ss_ar1_saz(serie)

\method{predict}{ss_ar1_saz}(object, n.ahead, ...)

\method{update}{ss_ar1_saz}(object, newseries, refit = FALSE, ...)
}
\arguments{
\item{serie}{serie para ajustar}

\item{object}{objeto com classes \code{c("ss_ar1_saz", "mod_eol")} contendo modelo}

\item{n.ahead}{número de passos à frente para previsão}

\item{...}{existe apenas para consistência com a genérica}

\item{newseries}{nova série com a qual atualizar o modelo}

\item{refit}{booleano indicando se o modelo deve ou não ser reajustado}
}
\value{
Objeto da classe mod_eol e subclasse \code{ss_ar1_saz}, uma lista de dois elementos:
\code{modelo} e \code{serie} contendo o modelo estimado e a série passada

\code{predict} retorna uma série temporal multivariada contendo valor esperado e desvio
padrao da previsão \code{n.ahead} passos à frente;

\code{update} retorna modelo com novos dados e, caso \code{refit == TRUE}, reajustado.
Contrário à função de estimação, \code{update} já retorna o objeto da classe \code{mod_eol};
}
\description{
Estimação e métodos de modelos da classe \code{ss_ar1_saz}

A previsão destes modelos é feita com alguns argumetnos opcionais já passados por padrão que
não podem ser modificados. Estes são: \code{se.fit = TRUE, filter = TRUE}. O primeiro retorna
além do valor previsto o desvio padrão associado, o segundo garante que são retornados os valores
advindos da distribuição preditiva e não de suavização. Considerando estes fatores, \code{...}
não pode conter \code{n.ahead} ou estes dois outros, ou então ocorrerá erro.
}
\details{
Esta especificação corresponde a um modelo em espaço de estados com um estado AR1 e outro de
sazonalidade correspondente à frequência de \code{serie}. O coeficiente autoregressivo será
estimado sob restrição de estacionariedade, assumindo valores necessariamente entre -1 e 1.

\bold{Update}:

A atualização de modelos \code{ss_ar1_saz} sempre vai checar se o modelo passado foi estimado
corretamente. Como modelos em espaçoo de estados dependem bastante de inicialização, às vezes não
dá para estimar direito. Nesses casos ele tenta reestimar o modelo independentemente de
\code{refit}
}
