% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod-ss_reg_din_pm.r
\name{modelos_ss_reg_din_pm}
\alias{modelos_ss_reg_din_pm}
\alias{ss_reg_din_pm}
\alias{predict.ss_reg_din_pm}
\alias{update.ss_reg_din_pm}
\title{Modelos \code{ss_reg_din_pm}}
\usage{
ss_reg_din_pm(serie, regdata, formula, vardin = FALSE, ...)

\method{predict}{ss_reg_din_pm}(object, newdata, n.ahead, ...)

\method{update}{ss_reg_din_pm}(object, newseries, newregdata, refit = FALSE, ...)
}
\arguments{
\item{serie}{série com sazonalidade a ser ajustada}

\item{regdata}{\code{data.frame}-like contendo variáveis explicativas}

\item{formula}{opcional, fórmula da regressão. Se for omitido, todas as variaveis em
\code{regdata} serão utilizadas}

\item{vardin}{booleano indicando se deve ser estimado modelo com heterocedasticidade.}

\item{...}{nao possui uso, existe apenas para consistencia com a generica}

\item{object}{objeto com classes \code{c("ss_reg_din_pm", "modprev")} contendo modelo}

\item{newdata}{\code{data.frame}-like contendo variéveis explicativas fora da amostra}

\item{n.ahead}{número de passos à frente para previsão. Este argumento não é necessario, caso não
seja informado a previsão sera feita tantos passos à frente quanto amostras em \code{newdata}}

\item{newseries}{nova série com a qual atualizar o modelo}

\item{newregdata}{\code{data.frame}-like contendo variáveis explicativas na nova amostra}

\item{refit}{booleano indicando se o modelo deve ou nao ser reajustado}
}
\value{
Objeto da classe \code{modprev} e subclasse \code{ss_reg_din_pm}, uma lista de dois
elementos: \code{modelo} e \code{serie} contendo o modelo estimado e a série passada

\code{predict} serie temporal multivariada contendo valor esperado e desvio padrao de
previsão \code{n.ahead} passos à frente;

\code{update} retorna modelo com novos dados e, caso \code{refit == TRUE}, reajustado.
Contrário à função de estimação, \code{update} já retorna o objeto da classe \code{modprev};
}
\description{
Estimação e métodos de modelos da classe \code{ss_reg_din_pm}

Todos os argumentos deste método de modelagem possuem a mesma definição dos argumentos homonimos
em \code{\link{ss_reg_din}}, portanto refira-se a esta pagina para mais detalhes. Do ponto de
vista do usuário, \code{ss_reg_din_pm} opera exatamente como \code{ss_reg_din}.

A previsão destes modelos é feita com alguns argumetnos opcionais já passados por padrão que
não podem ser modificados. Estes são: \code{se.fit = TRUE, filter = TRUE}. O primeiro retorna
além do valor previsto o desvio padrão associado, o segundo garante que são retornados os valores
advindos da distribuição preditiva e não de suavização. Considerando estes fatores, \code{...}
não pode conter \code{n.ahead} ou estes dois outros, ou então ocorrerá erro.

A atualização de modelos \code{ss_reg_din_pm} sempre vai checar se o modelo passado foi estimado
corretamente. Como modelos em espaçoo de estados dependem bastante de inicialização, às vezes não
dá para estimar direito. Nesses casos ele tenta reestimar o modelo independentemente de
\code{refit}
}
\details{
A diferença entre estes métodos consite na forma como a regressão é montada. Na versão padrão de
regressão dinâmica, a série temporal é tratada em sua forma univariada original. Isto pode levar
a problemas na previsão especialmente quando em horizontes mais longos (caso de remoção de viés
D1 ou mais).

\code{ss_reg_din_pm} manipula a série para um formato multivariado em no \code{R^s} em que cada
observação corresponde a uma sazonalidade da série original (no caso de dados horários, a série
univariada vira uma série multivariada de dimensão 24 em que cada observação é um dia completo).
O dado de variáveis explicativas \code{regdata} é quebrado em blocos de \code{s} linhas, sendo
cada bloco utilizado como matriz \code{Z_t} correspondente a cada observação multivariada.

Usualmente modelos multivariados com variáveis exógenas tem um vetor de variáveis explicativas e
uma matriz de coeficientes de regressão a ser estimada. Nesta configuração, cada dimensão da
variável dependente é uma função separada do mesmo conjunto de variáveis explicativas. Ao
formular o modelo do modo que \code{ss_reg_din_pm} faz, os coeficientes são compartilhados entre
todas a dimensões da variável explicativa, de modo que sua atualização é feita contemplando todas
as estações conjuntamente.

Este modelo foi concebido no contexto da correção de viés, especificamente para aplicação com
a regressão usando produto tensor de splines.

A opção \code{vardin}, neste caso, continua representando heterocedasticidade no sentido de que
cada estação possui uma variância própria, mas em termos mais precisos o modelo formulado é
homocedástico. Isto porque, ao se tornar multivariado, o choque \code{epsilon_t} passa a ser
vetorial com uma matriz de variâncias e covariâncias diagonal (na qual os elementos são
diferentes).
}
