% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod-vpar.r
\name{modelos_vpar}
\alias{modelos_vpar}
\alias{vpar}
\title{Modelos \code{vpar}}
\usage{
vpar(
  serie,
  s = frequency(serie),
  p = "auto",
  A12 = FALSE,
  max.p = 11,
  diag = FALSE,
  ...
)
}
\arguments{
\item{serie}{série para ajustar}

\item{s}{periodo de sazonalidade, caso \code{serie} nao seja uma serie temporal com sazonalidade}

\item{p}{vetor de comprimento igual a sazonalidade contendo as ordens de cada modelo periódico;
caso tenha comprimento menor que \code{s}, sera reciclado. Ver Detalhes}

\item{A12}{booleano indicando se a parcela anual deve ser incorporada à estimação}

\item{max.p}{vetor de comprimento igual a sazonalidade contendo as máximas ordens para
identificação automática; será reciclado se necessário. Ignorado se \code{p = "auto"}}

\item{diag}{booleano indicando o tipo de modelo a ser ajustado. Ver Detalhes}

\item{...}{nao possui uso, existe apenas para consistencia com a generica}
}
\value{
Objeto da classe \code{modprev} e subclasse \code{vpar}, uma lista de dois
elementos: \code{modelo} e \code{serie} contendo o modelo estimado e a série passada
}
\description{
Estimação e métodos de modelos da classe \code{vpar}

Assim como modelos PAR, os VPAR são necessariamente ajustados a séries temporais sazonais, com a
diferença de que estes são feitos para séries multivariadas. Em termos de identificação e uso dos
argumentos da função, todos funcionam exatamente tal qual aquelas de \code{\link{par}}, exceto
por \code{p}.
}
\details{
Modelos vetoriais periódicos autorregressivos, possivelmente com parcela anual, similares aos do
GEVAZP

Há apenas um argumento novo, \code{diag}, correspondendo a forma como o aspecto multivariado
é incorporado ao modelo. Caso \code{diag = TRUE}, o modelo e esitmado tal que suas matrizes
de parâmetros são bloco diagonais, isto é, cada série depende apenas de seus próprios lags.
Apenas o ruído permanece multivariado, considerando possíveis dependências entre séries.

Caso \code{diag = FALSE} as matrizes de parâmetros são consideradas cheias, introduzindo
dependência no sinal entre séries. Como este modelo tende a ficar sobrecarregado de parâmetros,
ele é estimado através do método MSGLasso realizando seleção de séries e então dos lags a serem
utilizados.

Assim, o argumento \code{p} funciona de forma diferente entre os modos diagonal ou não. No
primeiro ele se comporta tal qual em \code{\link{par}}, com uma adição. Caso seja informado um
escalar ou vetor de tamanho igual a sazonalidade de \code{serie}, a mesma informação será usada
para cada dimensão da série multivariada. Alternativamente pode ser fornecida uma lista de
\code{p}s e, então, cada elemento da lista será usado para uma dimensão da série. Quando se
estima o modelo não diagonal, \code{p} não tem uso pois a identificação dos lags será feita
automaticamente atráves do MSGLasso
}
