% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modprev-package.r
\docType{package}
\name{modprev-package}
\alias{modprev}
\alias{modprev-package}
\title{modprev: Interface Unificada Para Estimacao E Uso De Modelos}
\description{
Arcabouco unificado para estimacao, previsao, simulacao e etc de diversos tipos de modelos, incluindo aqueles com variaveis explicativas. Alem de wrappers para metodos tradicionais de modelos estatisticos, e implementada uma funcao para modelagem e previsao em horizonte rolante, facilitando testes com novos modelos. As funcoes do pacote operam em grande parte como wrappers em torno dos metodos especificos de cada tipo de modelo, visando apresentar uma interface comum para uso de todos eles, tal que a adicao de novos modelos seja simples.
}
\author{
\strong{Maintainer}: Lucas Khenayfis \email{lucaskhenayfis@gmail.com}

}
\keyword{internal}
