% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classe-mod_eol.r
\name{new_mod_eol}
\alias{new_mod_eol}
\title{Contrutor Interno De \code{mod_eol}}
\usage{
new_mod_eol(fit, serie, tipo)
}
\arguments{
\item{fit}{modelo estimado}

\item{serie}{serie para qual o modelo foi estimado}

\item{tipo}{string indicando espcificação do modelo}
}
\value{
Objeto da classe mod_eol e subclasse igual a \code{tipo}, uma lista de dois elementos:
\code{modelo} e \code{serie} contendo o modelo estimado e a série passada
}
\description{
Função interna, não deve ser chamada diretamente pelo usuário
}
