% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classe-modprevU.r
\name{new_modprevU}
\alias{new_modprevU}
\title{Contrutor Interno De \code{modprevU}}
\usage{
new_modprevU(fit, serie, tipo, atrs)
}
\arguments{
\item{fit}{modelo estimado}

\item{serie}{serie para qual o modelo foi estimado}

\item{tipo}{string indicando espcificação do modelo}

\item{atrs}{lista nomeada contendo atributos extras pertinentes ao modelo. Ver Detalhes}
}
\value{
Objeto da classe \code{modprevU} e subclasse igual a \code{tipo}, uma lista de dois
elementos: \code{modelo} e \code{serie} contendo o modelo estimado e a série passada.
Adicionalmente, se \code{atrs} for passada, um atributo "mod_atrs" contendo o argumento
}
\description{
Função interna, não deve ser chamada diretamente pelo usuário
}
\details{
\code{atrs} Existe para permitir que outras informacoes, nao necessariamente contidas no objeto
do modelo (como por exemplo a formula de regressao nos modelos de regressao dinamica), sejam
passadas adiante para os metodos de cada modelagem. A lista aqui passada sera adicionada ao
objeto \code{modprevU} de saida como um atributo chamado "mod_atrs".
}
