% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod-ss_reg_din.r
\name{parsedesloc}
\alias{parsedesloc}
\title{Deslocamento Do Array De Variancias}
\usage{
parsedesloc(serie, newseries, saz)
}
\arguments{
\item{serie}{a serie original do modelo}

\item{newseries}{nova serie para update}

\item{saz}{frequencia da heterocedasticidade}
}
\value{
inteiro indicando número de posicoes a deslocar por \code{\link{shift}}
}
\description{
Identifica a necessidade de deslocar o array H dependendo de em que instante cada serie comeca
}
\details{
Modelos com heterocedasticidade tem um array de variancias H, ao inves de uma matriz simples.
Isto significa que, dependendo de em que indice sazonal a serie original e nova comecam, pode ser
necessario deslocar o array um número de posicoes de modo que as variancias corretas estejam
associadas as novas observacoes. Por exemplo, se serie original tem freq = 4, H seria um array de
quatro posicoes ao longo da terceira dimensao. Se serie comecava em (XX, 3), as variancias em H
estao associadas aos indices 3, 4, 1, 2. Desta forma, se newseries comeca em (YY, 2), H deve ser
deslocado uma posicao a direita, de modo que o array resultante tenha, em ordem, as variacias
associadas aos indices 2, 3, 4, 1, o que e a ordem em que se encontram em newseries.

Tudo isso so faz sentido se assumirmos que tanto serie quanto newseries foram fornecidos
como objetos de serie temporal, de modo que e possivel identificar esse tipo de coisa.
Caso serie tenha sido um vetor, sera assumido que comeca no indice sazonal 1, o que pode
estar errado. Similarmente, se newseries for um vetor simples, sera assumido que comeca
no instante de tempo imediatamente apos o final da serie original.
}
