% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod-par.r
\name{perpacf}
\alias{perpacf}
\alias{percacf}
\title{Autocorrelação Parcial/Condicional Periódica}
\usage{
perpacf(serie, m, lag.max = 6, plot = FALSE)

percacf(serie, medias, m, lag.max = 6, plot = FALSE)
}
\arguments{
\item{serie}{serie temporal da qual calcular autocorrelações}

\item{m}{inteiro indicando o periodo da sazonalidade para o qual calcular autocorrelacoes}

\item{lag.max}{inteiro indicando o maior lag para calcular a autocorrelação, limitado em 11}

\item{plot}{booleano indicando se deve ser feito o plot das autocorrelações calculadas}

\item{medias}{serie temporal com mesmas características de \code{serie} contendo as medias moveis}
}
\description{
Calcula e opcionalmente plota a ACF parcial ou condicional de uma determinada série sazonal
}
