% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizacao.r
\name{plot.mod_eol}
\alias{plot.mod_eol}
\title{Plot De Objetos \code{mod_eol}}
\usage{
\method{plot}{mod_eol}(x, ...)
}
\arguments{
\item{...}{demais parametros repassados para \code{plot}}

\item{fit}{objeto para plotar}
}
\value{
plot da serie e ajuste realizado
}
\description{
Metodo para plotar ajustes \code{mod_eol}
}
