% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizacao.r
\name{plot.modprev}
\alias{plot.modprev}
\title{Plot De Objetos \code{modprev}}
\usage{
\method{plot}{modprev}(x, ...)
}
\arguments{
\item{x}{objeto para plotar}

\item{...}{demais parametros repassados para \code{plot}}
}
\value{
plot da serie e ajuste realizado
}
\description{
Metodo para plotar ajustes \code{modprev}
}
