% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimamodelo.r
\name{predict.modprev}
\alias{predict.modprev}
\title{Previsão De Modelos \code{modprev}}
\usage{
\method{predict}{modprev}(object, n.ahead, ...)
}
\arguments{
\item{object}{modelo ajustado através de \code{\link{estimamodelo}}}

\item{n.ahead}{número de passos à frente para prever}

\item{...}{Opcionalmente, pode ser passado o argumento \code{newdata} \code{data.frame}-like
contendo variaveis explicativas fora da amostra para modelos que necessitem}
}
\value{
série temporal multivariada contendo a previsão e o desvio padrão associado para os
passos de tempo \code{1:n.ahead}
}
\description{
Wrapper para previsão de modelos ajustados por \code{\link{estimamodelo}}
}
\details{
A previsão destes modelos funciona da forma mais padrão, com \code{object} contendo um modelo
estimado por \code{\link{estimamodelo}} e \code{n.ahead} um inteiro indicando número de passos à
frente para prever. Em modelos de séries temporais simples, isto é tudo.

\bold{Modelos com variáveis explicativas:}

No caso de modelos com variáveis explicativas deve ser fornecido um parâmetro \code{newdata} na
forma de code{data.frame}-like contendo as variáveis necessárias para o modelo. Usualmente a
previsão será feita tantos passos à frente quanto há observações em \code{newdata}, porém se
\code{n.ahead} for fornecido será usado com precedência sobre o número de observações novas.

\bold{Modelos periódicos:}

Nos casos de modelos periódicos que necessitam variaveis explicativas, alguns cuidados devem ser
tomados. O usuario informara um argumento \code{newdata} contendo um \code{data.frame}-like unico
contendo todas as variaveis explicativas fora da amostra. Se assume que as linhas neste dado
correspondem as variaveis explicativas nos tempos apos o ultimo da serie, cronologicamente.

Isto significa que, se a serie orignal era mensal terminando em junho/2020, o programa assume que
\code{newdata} tem, na primeira linha, as variaveis explicativas para julho/2020, na segunda,
para ago/2020 e assim por diante. Esta suposicao se sustenta no fato de que o pacote
\code{modprev} foi feito para modelagem de  series temporais em principio.

Alternativamente, o usuario pode passar \code{newdata} como uma lista de \code{data.frame}-likes;
neste caso se assume que cada elemento da lista corresponde a uma estacao do dado, em ordem.
Observe que, se a serie comeca em maio, entao MAIO CORRESPONDE A PRIMEIRA ESTACAO.
}
\examples{

mod <- estimamodelo(AirPassengers, tipo = "sarima")
pred <- predict(mod, n.ahead = 24)

# em modelos de regressao, deve ser passada a variavel explicativa via newdata

serie <- window(datregdin$obs, 1, 100)
varex <- datregdin$varex[1:100, , drop = FALSE]
mod_regdin <- estimamodelo(serie, "ss_reg_din", regdata = varex)

newdata <- datregdin$varex[101:130, , drop = FALSE]
pred <- predict(mod_regdin, newdata = newdata)

}
\seealso{
Other Metodos modprev: 
\code{\link{estimamodelo}()},
\code{\link{update.modprev}()}
}
\concept{Metodos modprev}
