% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classe-modprevP.r
\name{predict.modprevP}
\alias{predict.modprevP}
\title{Previsao De Modelos Periodicos}
\usage{
\method{predict}{modprevP}(object, n.ahead, ...)
}
\arguments{
\item{object}{modelo periodico com o qual realizar a previsao}

\item{n.ahead}{número de passos à frente para previsão. Este argumento não é necessario, caso não
seja informado a previsão sera feita tantos passos à frente quanto amostras em \code{newdata}}

\item{...}{Opcionalmente, pode ser passado o argumento \code{newdata} \code{data.frame}-like
contendo variaveis explicativas fora da amostra para modelos que necessitem}
}
\value{
série temporal multivariada contendo a previsão e o desvio padrão associado para os
passos de tempo \code{1:n.ahead}
}
\description{
Wrapper para previsao de cada modelo individual e reorganizacao em serie unica
}
\details{
Nos casos de modelos que necessitam variaveis explicativas, alguns cuidados devem ser tomados. O
usuario informara um argumento \code{newdata} contendo um \code{data.frame}-like unico contendo
todas as variaveis explicativas fora da amostra. Se assume que as linhas neste dado correspondem
as variaveis explicativas nos tempos apos o ultimo da serie, cronologicamente.

Isto significa que, se a serie orignal era mensal terminando em junho/2020, o programa assume que
\code{newdata} tem, na primeira linha, as variaveis explicativas para julho/2020, na segunda,
para ago/2020 e assim por diante. Esta suposicao se sustenta no fato de que o pacote
\code{modprev} foi feito para modelagem de  series temporais em principio.

Alternativamente, o usuario pode passar \code{newdata} como uma lista de \code{data.frame}-likes;
neste caso se assume que cada elemento da lista corresponde a uma estacao do dado, em ordem.
Observe que, se a serie comeca em maio, entao MAIO CORRESPONDE A PRIMEIRA ESTACAO.
}
