% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classe-modprevS.r
\name{predict.modprevS}
\alias{predict.modprevS}
\title{\code{predict} De Modelos Com Shapesifting}
\usage{
\method{predict}{modprevS}(object, n.ahead, newdata_list, feedback_fun = NULL, ...)
}
\arguments{
\item{object}{objeto com classe \code{"modprevS"} contendo modelo}

\item{n.ahead}{número de passos à frente para previsão}

\item{newdata_list}{lista nomeada contendo os dados a serem usados na extracao via pipes}

\item{feedback_fun}{função opcional para atualizar \code{newdata_list} a cada passo de previsão}

\item{...}{Para \code{predict}, demais argumentos passados para a função de previsão do modelo interno}
}
\value{
\code{predict} retorna uma série temporal multivariada contendo valor esperado e desvio
padrão da previsão \code{n.ahead} passos à frente
}
\description{
Wrapper de previsao para modelos com shapeshifting
}
\details{
\bold{FUNCIONALIDADE EM TESTE - SUJEITO A MUDANCAS FUTURAS}

O argumento \code{feedback_fun} controla o modo de previsao sendo empregado. Caso nao seja fornecido,
realiza-se previsao single-shot com o maximo possivel de dados utilizaveis. Se for fornecido, se
aplica previsao recursiva, na qual cada previsao unitaria é feita e realimentada aos dados
\code{newdata_list} utilizando a funcao \code{feedback_fun}. \code{feedback_fun}, se fornecido, deve ser uma
funcao que recebe como primeiro argumento um vetor com as previsoes feitas ate o momento (apenas
valores previstos, sem desvios padrao) e como segundo argumento a \code{newdata_list} atual. A funcao
deve retornar uma lista nomeada com os novos valores a serem atualizados em \code{newdata_list} para o
proximo passo de previsao.
}
