% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classe-jm-config.r
\name{print.jm_config}
\alias{print.jm_config}
\title{Print Method for jm_config}
\usage{
\method{print}{jm_config}(x, ...)
}
\arguments{
\item{x}{jm_config object}

\item{...}{Additional arguments (ignored)}
}
\value{
Invisible x
}
\description{
Print Method for jm_config
}
