% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-registry.r
\name{print.model_registry}
\alias{print.model_registry}
\title{Print Method for model_registry}
\usage{
\method{print}{model_registry}(x, ...)
}
\arguments{
\item{x}{Object of class \code{"model_registry"}.}

\item{...}{Additional arguments (unused).}
}
\value{
The input object \code{x}, invisibly.
}
\description{
Displays registered models in a readable format with indicators for models requiring
regression data.
}
