% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-spec.r
\name{print.model_spec}
\alias{print.model_spec}
\title{Print Method for model_spec}
\usage{
\method{print}{model_spec}(x, ...)
}
\arguments{
\item{x}{Object of class \code{"model_spec"}.}

\item{...}{Additional arguments (unused).}
}
\value{
The input object \code{x}, invisibly.
}
\description{
Displays model specification information in a readable format.
}
