% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{scale_by_season}
\alias{scale_by_season}
\title{Normaliza Serie Sazonalmente}
\usage{
scale_by_season(serie, est = "n", truncdat = -1, truncpar = -1)
}
\arguments{
\item{serie}{serie temporal com sazonalidade para normalizar}

\item{est}{string indicando qual estimador utilizar para desvio padrao
- "n" estimador normalizando somatorios por n
- "n-1" estimador normalizando somatorios por (n - 1)}

\item{truncdat}{inteiro indicando o numero de casas decimais para arredondamento do dado
fornecido. Padrao -1 = nao arredondar}

\item{truncpar}{inteiro indicando o numero de casas decimais para arredondamento das medias e sd
para normalizacao. Padrao -1 = nao arredondar}
}
\value{
\link{matriz numerica} matriz contendo dat normalizado
}
\description{
Funcao interna para padronizar uma serie por periodo de sazonalidade
}
