% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{sd2}
\alias{sd2}
\title{Calculo do desvio padrao normalizado por 1/n}
\usage{
sd2(...)
}
\arguments{
\item{vec}{\link{vetor numerico} serie cujo desvio padrao se deseja calcular}
}
\value{
\link{escalar numerico} desvio padrao calculado
}
\description{
Funcao alternativa para calcular desvio padrao atraves do estimador "n"
}
