% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{set_sim_seed}
\alias{set_sim_seed}
\title{Configura Seed Para Simulacao}
\usage{
set_sim_seed(seed)
}
\arguments{
\item{seed}{valor da semente ou \code{NULL} para nao definir}
}
\description{
Utilitario para definir semente de geracao aleatoria em simulacoes
}
\keyword{internal}
