% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{shift}
\alias{shift}
\title{Desloca Vetor De \code{i} Posicoes}
\usage{
shift(vec, i)
}
\arguments{
\item{i}{posicoes para deslocar -- se negativo, para a esquerda, positivo para a direita}

\item{v}{vetor a ser manipulado}
}
\value{
vetor \code{v} com elementos deslocados \code{i} posicoes
}
\description{
Desloca os valores de um vetor mantendo o mesmo comprimento original
}
