% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{sim_ts}
\alias{sim_ts}
\title{Formata Saida De Simulacao}
\usage{
sim_ts(sims, serie)
}
\arguments{
\item{sims}{matriz numerica (\code{n.ahead x nsim}) ou vetor numerico (\code{n.ahead}) com os
valores simulados}

\item{serie}{a serie ajustada (um \code{ts} ou numerico coercivel via \code{as.ts})}
}
\value{
\code{ts} multivariado com \code{nrow(sims)} linhas e \code{ncol(sims)} colunas,
nomeadas \code{sim_1..sim_ncol(sims)}, iniciando um passo apos o fim de \code{serie}
}
\description{
Padroniza a saida de simulacoes de modelos como um \code{ts} multivariado, com colunas
nomeadas e sistema de tempo continuando a serie ajustada
}
\details{
Espelha a logica de posicionamento de \code{predict.modprevP} (\code{R/classe-modprevP.r}),
que calcula o proximo instante como \code{tp1 = aux_tsp[2] + 1 / aux_tsp[3]}
}
\keyword{internal}
