% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimamodelo.r
\name{simulate.modprev}
\alias{simulate.modprev}
\title{Simulacao De Modelos \code{modprev}}
\usage{
\method{simulate}{modprev}(object, nsim = 1, seed = NULL, n.ahead, ...)
}
\arguments{
\item{object}{modelo ajustado através de \code{\link{estimamodelo}}}

\item{nsim}{número de simulações a serem geradas}

\item{seed}{opcionalmente, semente para geração das simulações}

\item{n.ahead}{número de passos à frente para simular}

\item{...}{Opcionalmente, pode ser passado o argumento \code{newdata} \code{data.frame}-like
contendo variaveis explicativas fora da amostra para modelos que necessitem}
}
\value{
série temporal multivariada contendo \code{nsim} simulações para os passos de tempo
\code{1:n.ahead}
}
\description{
Wrapper para simulacao de modelos ajustados por \code{\link{estimamodelo}}
}
\seealso{
Other Metodos modprev: 
\code{\link{estimamodelo}()},
\code{\link{predict.modprev}()},
\code{\link{update.modprev}()}
}
\concept{Metodos modprev}
