% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classe-modprevS.r
\name{simulate.modprevS}
\alias{simulate.modprevS}
\title{Simulacao De Modelos \code{modprevS}}
\usage{
\method{simulate}{modprevS}(object, nsim = 1, seed = NULL, n.ahead, ...)
}
\arguments{
\item{object}{modelo ajustado através de \code{\link{estimamodelo_S}}}

\item{nsim}{número de simulações a serem geradas}

\item{seed}{opcionalmente, semente para geração das simulações}

\item{n.ahead}{número de passos à frente para simular}

\item{...}{demais argumentos, nao utilizados}
}
\value{
este metodo sempre gera erro, pois simulacao de modelos \code{modprevS} nao e suportada
}
\description{
Wrapper para simulacao de modelos com shapeshifting
}
