% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{split_seasonal_regdata}
\alias{split_seasonal_regdata}
\title{Divide Dados De Variaveis Explicativas Por Estacao}
\usage{
split_seasonal_regdata(regdata, seasons)
}
\arguments{
\item{regdata}{\code{data.frame}, matriz ou outro objeto aceito por \code{split()}}

\item{seasons}{fator de estacoes conforme construido por \code{\link{split_seasonal}}}
}
\value{
lista nomeada resultante de \code{split(regdata, seasons)}, com nomes iguais aos
niveis de \code{seasons}
}
\description{
Divide um \code{data.frame}-like de variaveis explicativas em sublistas por estacao,
usando o fator de estacoes produzido por \code{split_seasonal}
}
\details{
Esta funcao centraliza a chamada \code{split(regdata, seasons)} utilizada nos modelos
periodicos. Nao trata o caso em que \code{regdata} e uma lista -- esse caso permanece
inline nos pontos de chamada.
}
\keyword{internal}
