% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelos.r
\name{update.mod_eol}
\alias{update.mod_eol}
\title{Atualizacao De Modelos mod_eol}
\usage{
\method{update}{mod_eol}(object, newseries, refit = FALSE, ...)
}
\arguments{
\item{newseries}{nova serie para associar ao modelo}

\item{refit}{booleano indicando se o modelo deve ser reajustado}

\item{...}{demais parâmetros passados para as updates específicas. Ver Detalhes}

\item{fit}{modelo ajustado atraves de estimamodelo}
}
\value{
modelo com novos dados e possivelmente reajustado
}
\description{
Wrapper para atualizar e possivelmete reajustar modelos \code{mod_eol}
}
\details{
O padrão desta função é simplesmente substituir \code{newseries} no modelo ajustado \code{fit},
isto é, mantendo todos os hiperparâmetros estimados originalmente. Através do argumento
\code{refit} é possível realizar o reajuste do modelo para a nova série.

No caso de modelos com variaveis explicativas deve ser fornecido um parametro \code{newregdata}
na forma de uma matriz ou data.frame contendo apenas as colunas com variáveis a serem utilizadas.
Se houver apenas uma variável explicativa, pode ser passada como um vetor ou série temporal.
}
\examples{

serie1 <- window(datregdin[[1]], 1, 300)
serie2 <- window(datregdin[[1]], 501, 800)

mod_orig  <- estimamodelo(serie1, tipo = "sarima")
mod_upd   <- update(mod_orig, serie2, refit = FALSE)
mod_refit <- update(mod_orig, serie2, refit = TRUE)

\dontrun{
# comparando os casos
coef(mod_orig$modelo)
coef(mod_upd$modelo)
coef(mod_refit$modelo)
}

}
