% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classe-modprevP.r
\name{update.modprevP}
\alias{update.modprevP}
\title{Update De Modelos Periodicos}
\usage{
\method{update}{modprevP}(object, newseries, refit = FALSE, ...)
}
\arguments{
\item{object}{modelo ajustado atraves de \code{estimamodelo_U}}

\item{newseries}{nova serie para associar ao modelo}

\item{refit}{booleano indicando se o modelo deve ser reajustado}

\item{...}{Opcionalmente, pode ser passado o \code{newregdata}, um \code{data.frame}-like
contendo variaveis explicativas pareadas com \code{newseries} para modelos que as necessitem}
}
\value{
modelo com novos dados, possivelmente reajustado
}
\description{
Wrapper para atualizar e possivelmente reajustar modelos periodicos
}
