% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classe-modprevU.r
\name{update.modprevU}
\alias{update.modprevU}
\title{Atualizacao De Modelos \code{modprevU}}
\usage{
\method{update}{modprevU}(object, newseries, refit = FALSE, ...)
}
\arguments{
\item{object}{modelo ajustado atraves de \code{estimamodelo_U}}

\item{newseries}{nova serie para associar ao modelo}

\item{refit}{booleano indicando se o modelo deve ser reajustado}

\item{...}{Opcionalmente, pode ser passado o \code{newregdata}, um \code{data.frame}-like
contendo variaveis explicativas pareadas com \code{newseries} para modelos que as necessitem}
}
\value{
modelo com novos dados, possivelmente reajustado
}
\description{
Wrapper para atualizar e possivelmete reajustar modelos \code{modprevU}
}
\seealso{
Other Metodos modprevU: 
\code{\link{estimamodelo_U}()},
\code{\link{predict.modprevU}()}
}
\concept{Metodos modprevU}
