% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod-ss_reg_din.r
\name{update_funs}
\alias{update_funs}
\alias{updH_homoc}
\alias{updH_heter_trig}
\alias{updQ_stat}
\alias{updQ_din}
\title{Auxiliares Para Estimacao}
\usage{
updH_homoc(par, mod, ...)

updH_heter_trig(par, mod, saz, ...)

updQ_stat(par, mod, ...)

updQ_din(par, mod, ...)
}
\arguments{
\item{par}{vetor completo de hiperparametros. Ver Detalhes}

\item{mod}{o modelo a ser atualizado}

\item{...}{sem uso, apenas para consistencia de argumentos entre funcoes}

\item{saz}{sazonalidade da serie modelada}
}
\value{
\code{mod} com a matriz modificada
}
\description{
Funcoes de atualizacao das matrizes H e Q num modelo espaco de estados
}
\details{
Cada funcao \code{updX_*} atualiza a matriz \code{X} do sistema por uma abordagem diferente.
Independentemente do numero de hiperparametros necessarios, cada funcao deve receber o vetor
completo e extrai dele os que sao pertinentes para si. Por padrao, para facilitar este processo,
os hiperparametros associado a matriz H sempre estao primeiro e os a Q por ultimo.

Por exemplo, \code{updH_heter_trig} assume que as variancias sazonais sao geradas por uma funcao
trigonometrica de um harmonico definida em termos de dois hiperparametros, os dois primeiros
elementos do vetor \code{par}. Os demais, quantos quer que sejam, sao relativos a Q.
}
