% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-validation.r
\name{validate_control_split}
\alias{validate_control_split}
\title{Validate Split Mode Control Parameters}
\usage{
validate_control_split(control_list, serie_length)
}
\arguments{
\item{control_list}{Named list of validation control parameters}

\item{serie_length}{Integer length of the series being modeled}
}
\value{
The input control_list (invisibly) if valid, otherwise raises error
}
\description{
Checks that control list contains required 'oob' parameter for train/test split
}
