% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/janelamovel.r
\name{verbose_fun}
\alias{verbose_fun}
\title{Selecao Da Funcao De Log}
\usage{
verbose_fun(verbose)
}
\arguments{
\item{verbose}{inteiro indicando o nivel de detalhe a ser informado:
0 = nada, 1: toda vez que reajusta modelo, 2: todo horizonte de previsao e reajuste}
}
\value{
funcao para uso interno na janela movel que printa o estado de execucao
}
\description{
Retorna uma funcao para prints informativos durante a execucao
}
