% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{datregdin}
\alias{datregdin}
\title{Dado Exemplo De Velocidade Do Vento}
\format{
lista de dois elementos:
\describe{
\item{obs}{série temporal de 200 observações do observado dummy}
\item{prev}{\code{data.frame} de três colunas contendo as variáveis explicativas}
}
}
\usage{
datregdin
}
\description{
Dado dummy para exemplificação dos modelos de regressão dinâmica.
}
\keyword{datasets}
