% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.r
\name{deltats}
\alias{deltats}
\title{Índice De Tempo Por Delta Em Série Temporal}
\usage{
deltats(ini, fim, delta, freq)
}
\description{
Calcula índice no sistema de tempo de um ts especificada para um dado delta
}
