% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classe-mod_eol.r
\name{estimamodelo}
\alias{estimamodelo}
\title{Ajuste De Modelos}
\usage{
estimamodelo(serie, tipo, ...)
}
\arguments{
\item{serie}{série para ajustar}

\item{tipo}{tipo de modelo a ser ajustado. Ver Detalhes}

\item{...}{demais parâmetros passados para as funções de fit específicas de cada modelo}
}
\value{
objeto da classe mod_eol contendo modelo (classe dependente do modelo ajustato), serie
ajustada e, caso \code{out_sample > 0}, a parte reservada para comparação
}
\description{
Wrapper de estimação para múltiplos modelos com interface e saída unificadas
}
\details{
Esta função facilita a estimação de diversos tipos de modelos com uma única interface e, mais
importante ainda, única estrutura de saída. O tipo de modelo estimado para a série passada
através de \code{serie} é selecionado através do argumento \code{tipo}, podendo ser um de

\describe{
\item{\code{sarima}}{SARIMA(p, d, q)(P, D, Q)}
\item{\code{ss_ar1_saz}}{Espaço de estados composto por processo AR(1) + Sazonalidade}
\item{\code{ss_reg_din}}{Regressão univariada dinâmica}
}

Deve ser notado que no caso dos modelos com sazonalidade, o argumento \code{serie} \emph{DEVE SER
UM OBJETO SERIE TEMPORAL COM PERIODO ESPECIFICADO}. Isto e necessario para que a função possa
automaticamente lidar com diversos tipos de séries sem a necessidade de demais argumentos.

No caso de modelos com variaveis explicativas deve ser fornecido um parametro \code{regdata} na
forma de uma matriz ou data.frame contendo apenas as colunas com variáveis a serem utilizadas. Se
houver apenas uma variável explicativa, pode ser passada como um vetor ou série temporal.
}
\examples{

# ajustando tipo SARIMA
mod_sarima <- estimamodelo(AirPassengers, tipo = "sarima")

# caso a serie nao possua sazonalidade explicita, o modelo sera ajustado sem isso
ss <- arima.sim(200, model = list(ar = .8))
mod_sarima_semsazo <- estimamodelo(ss, tipo = "sarima")

\dontrun{
# estima so um AR(1) sem sazonalidade
coef(mod_sarima_semsazo)
}

# ajustando uma regressao dinamica (com dado dummy interno do pacote)
serie <- window(datregdin[[1]], 1, 100)
varex <- window(datregdin[[2]], 1, 100)
mod_regdin <- estimamodelo(serie, regdata = varex, tipo = "reg_din")

}
\seealso{
Other metodos_mod_eol: 
\code{\link{predict.mod_eol}()},
\code{\link{update.mod_eol}()}
}
\concept{metodos_mod_eol}
