% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod-ss_reg_din.r
\name{fitSSM2}
\alias{fitSSM2}
\title{Estimacao De SSM Com lambda}
\usage{
fitSSM2(model, inits, updatefn, checkfn, update_args = NULL, lambda = 0, ...)
}
\arguments{
\item{model}{Model object of class \code{SSModel}.}

\item{inits}{Initial values for \code{\link{optim}}.}

\item{updatefn}{User defined function which updates the model given the
parameters. Must be of form \code{updatefn(pars, model, ...)},
where \code{...} correspond to optional additional arguments.
Function should return the original model with updated parameters.
See details for description of the default \code{updatefn}.}

\item{checkfn}{Optional function of form \code{checkfn(model)} for checking
the validity of the model. Should return \code{TRUE} if the model is valid,
and \code{FALSE} otherwise. See details.}

\item{update_args}{Optional list containing additional arguments to \code{updatefn}.}

\item{lambda}{penality coefficient}

\item{...}{Further arguments for functions \code{optim} and
\code{logLik.SSModel}, such as \code{nsim = 1000}, \code{marginal = TRUE},
and \code{method = "BFGS"}.}
}
\description{
Funcao que faz a estimacao de modelos em espaco de estado com penalizacao da matriz Q
}
\details{
\code{fitSSM2} e uma copia de \code{\link[KFAS]{fitSSM}}, com uma simples modificacao: a funcao
objetivo agora e uma ponderacao entre a verossimilhanca e uma penalizacao da variacao total da
matriz Q. Isto e feito no contexto de regressoes dinamicas com o intuito de controlar o quao
variaveis sao os regressores.

Para maiores detalhes, ver \code{\link[KFAS]{fitSSM}}, pois todo o restante de interface e saidas
e igual
}
