% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod-ss_reg_din.r
\name{fitSSM2}
\alias{fitSSM2}
\title{Estimacao De SSM Com lambda}
\usage{
fitSSM2(model, inits, updatefn, checkfn, update_args = NULL, lambda = 0, ...)
}
\description{
Funcao que faz a estimacao de modelos em espaco de estado com penalizacao da matriz Q
}
\details{
\code{fitSSM2} e uma copia de \code{\link[KFAS]{fitSSM}}, com uma simples modificacao: a funcao
objetivo agora e uma ponderacao entre a verossimilhanca e uma penalizacao da variacao total da
matriz Q. Isto e feito no contexto de regressoes dinamicas com o intuito de controlar o quao
variaveis sao os regressores.

Para maiores detalhes, ver \code{\link[KFAS]{fitSSM}}, pois todo o restante de interface e saidas
e igual
}
