% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-registry.r
\name{is_registered}
\alias{is_registered}
\title{Check if Model is Registered}
\usage{
is_registered(tipo)
}
\arguments{
\item{tipo}{Character string identifying the model type (e.g., \code{"sarima"}, \code{"reg_lin"}).}
}
\value{
Logical value: \code{TRUE} if the model is registered, \code{FALSE} otherwise.
}
\description{
Tests whether a model type exists in the global model registry.
}
\examples{
is_registered("sarima")

is_registered("nonexistent_model")

dummy_fit <- function(serie, ...) {
    new_modprevU(list(data = serie), serie, "test", list())
}
dummy_predict <- function(object, n.ahead, ...) ts(rep(0, n.ahead))
dummy_update <- function(object, newseries, refit = FALSE, ...) object

register_model("test", dummy_fit, dummy_predict, dummy_update)
is_registered("test")

modprev:::.remove_model("test")

}
\seealso{
\code{\link[=register_model]{register_model()}}, \code{\link[=get_registry]{get_registry()}}

Other model_registry: 
\code{\link{get_model}()},
\code{\link{get_registry}()},
\code{\link{list_models}()},
\code{\link{model_spec}()},
\code{\link{register_model}()}
}
\concept{model_registry}
