% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/janelamovel.r
\name{janelamovel}
\alias{janelamovel}
\title{Previsão Em Horizonte Rolante}
\usage{
janelamovel(
  serie,
  tipo,
  janela,
  passo = 1L,
  n.ahead = 1L,
  refit.cada = NA,
  verbose = 0,
  ...
)
}
\arguments{
\item{serie}{serie temporal pela qual passar a janela movel}

\item{tipo}{tipo de modelo a ser ajustado, caso \code{objeto} seja uma serie temporal}

\item{n.ahead}{numero de passos a frente para prever a cada passo}

\item{refit.cada}{escalar ou vetor inteiro. Se escalar, reajusta o modelo a cada
\code{refit.cada} observacoes. Se vetor, reajusta apos cada indice de \code{refit.cada}}

\item{verbose}{Escalar inteiro indicando quanta informacao a ser emitida durante rodada.
0 = nenhuma, 1: toda vez que reajusta modelo, 2: todo horizonte de previsao e reajuste}

\item{...}{demais parametros pertinentes a cada tipo específico de modelagem}

\item{largura}{numero de observacoes na janela movel}
}
\value{
lista contendo previsoes de 1 a n.ahead passos a frente para cada janela
}
\description{
Função para realizar previsões e reajustes em janela móvel
}
\details{
Para cada janela de tamanho \code{largura} na \code{serie} fornecida será realizada uma previsão
\code{n.ahead} passos à frente, retornadas em uma lista. Isto significa que a última previsão
realizada cobrirá até \code{n.ahead} além da última observação em \code{serie}.

O comportamento dessa função é mais fortemente impactado por \code{refit.cada}. Através deste
argumento é possível indicar observações ou intervalos de tempo nos quais o modelo será
reestimado. Nas demais janelas será feita apenas a atualização das informações.

Caso o modelo escolhido necessite de variáveis explicativas, é necessário que o argumento
\code{regdata} seja passado na forma de uma matriz ou data.frame contendo apenas as colunas com
variáveis a serem utilizadas. Se houver apenas uma variável explicativa, pode ser passada como um
vetor ou série temporal.

Este argumento \emph{DEVE CONTER AS VARIÁVEIS EXPLICATIVAS CORRESPONDENTES A TODAS AS OBSERVAÇÕES
DA SÉRIE MAIS \code{n.ahead} À FRENTE}. A primeira parte dessa restrição é natural, pois são
necessárias as variáveis explicativas insample para ajustes do modelo. As observações
\code{n.ahead} passos à frente do final da série são necessárias apenas para a previsão das
últimas janelas.
}
\examples{

\dontrun{
serie <- ts(window(datregdin[[1]], 1, 250))
jm <- janelamovel(serie, "ss_ar1_saz", 200, refit.cada = 10)

serie <- ts(window(datregdin[[1]], 1, 720), freq = 48)
jm <- janelamovel(serie, "ss_ar1_saz", 480, refit.cada = 48)

serie <- ts(window(datregdin[[1]], 1, 250))
varex <- datregdin[1:255, 2, drop = FALSE]
jm <- janelamovel(serie, "ss_reg_din", 200, 5L, refit.cada = 10, regdata = varex)

# como esta funcao pode demorar um tempo para rodar (especialmente com multiplos refits), pode
# ser conveniente mandar o log para um arquivo
sink("log_janelamovel.txt")
jm <- janelamovel(serie, "ss_ar1_saz", 200, refit.cada = 10, verbose = 2)
sink()
}

}
