% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/janelamovel.r
\name{janelamovel}
\alias{janelamovel}
\title{Previsão Em Horizonte Rolante}
\usage{
janelamovel(
  serie,
  tipo,
  janela,
  passo = 1L,
  n.ahead = 1L,
  refit.cada = NA,
  verbose = 0,
  ...
)
}
\arguments{
\item{serie}{serie temporal pela qual passar a janela movel}

\item{tipo}{tipo de modelo a ser ajustado. Ver \code{\link{estimamodelo}}.}

\item{janela}{especificação da janela móvel ou expansível}

\item{passo}{inteiro de saltos temporais entre cada janela. Ver Detalhes}

\item{n.ahead}{número de passos à frente para prever a cada passo}

\item{refit.cada}{escalar indicando de quantas em quantas observacoes o modelo deve ser
reajustado}

\item{verbose}{Escalar inteiro indicando quanta informacao a ser emitida durante rodada.
0 = nenhuma, 1: toda vez que reajusta modelo, 2: todo horizonte de previsão e reajuste}

\item{...}{demais argumentos pertinentes a estimação de cada \code{tipo}. Veja
\code{\link{estimamodelo}} para mais detalhes}
}
\value{
lista contendo previsoes de 1 a n.ahead passos à frente para cada janela
}
\description{
Função para realizar previsões e reajustes em janela móvel
}
\details{
Os argumentos \code{serie} e \code{tipo} tem exatamente o mesmo efeito daqueles descritos em
\code{\link{estimamodelo}}.

\code{janela} pode ser ou um escalar ou um vetor de dois elementos inteiros. No primeiro caso,
entende-se que deverao ser rodadas janelas de tamanho \code{janela}. Quando este argumento e um
vetor entende-se que todas as janelas devem iniciar no instante de tempo \code{janela[1]} com
largura inicial \code{janela[2]}. Essencialmente estas duas formas permitem a execucao em janela
rolante ou expansivel, respectivamente.

O argumento \code{passo} permite especificar quantas novas observacoes sao incorporadas entre
janelas adjacentes. No caso de janela rolante, cada janela incorpora \code{passos} novos pontos e
abandona os \code{passos} valores mais antigos, de modo que todas tem o mesmo tamanho; janelas
expansiveis vao simplesmente agregando novas observacoes. Deve ser observado que, no caso de
janelas rolantes, sempre havera uma janela contendo o final da serie, mesmo que passos seja um
número tal que da penultima para a ultima janela ocorra um intervalo menor que \code{passos}.

O comportamento dessa função é mais fortemente impactado por \code{refit.cada}. Através deste
argumento é possível indicar intervalos de tempo nos quais o modelo será reestimado. Nas demais
janelas será feita apenas a atualização das informações. Caso \code{refit.cada = NA} (o padrão),
o modelo só será ajustado uma vez e atualizado a cada nova janela.

\bold{Modelos com variáveis explicativas:}

Caso o modelo escolhido use de variáveis explicativas, é necessário que o \code{regdata} seja
passado na forma de um \code{data.frame}-like contendo as variáveis a serem utilizadas.

Este argumento \emph{DEVE CONTER AS VARIÁVEIS EXPLICATIVAS CORRESPONDENTES A TODAS AS OBSERVAÇÕES
DA SÉRIE MAIS \code{n.ahead} À FRENTE}. A primeira parte dessa restrição é natural, pois são
necessárias as variáveis explicativas insample para ajustes do modelo. As observações
\code{n.ahead} passos à frente do final da série são necessárias apenas para a previsão das
últimas janelas.
}
\examples{

# janela rolante de cinco anos, prevendo um ano a frente e rolando de seis em seis meses
jm_sarima <- janelamovel(AirPassengers, "sarima", 60, passo = 6, n.ahead = 12)

# mesma configuracao, porem janela extensivel (inicio fixo no primeiro mes)
jm_sarima <- janelamovel(AirPassengers, "sarima", c(1, 60), passo = 6, n.ahead = 12)

# VARIAVEIS EXPLICATIVAS -----------------------------------------

# serie deve conter n.ahead menos observacoes que a variavel explicativa
serie <- window(datregdin$obs, 1, 190)
varex <- datregdin$varex
jm_regdin <- janelamovel(serie, "ss_reg_din", 100, passo = 10, n.ahead = 5, regdata = varex)

# MODO VERBOSE ---------------------------------------------------

\dontrun{
jm_sarima <- janelamovel(AirPassengers, "sarima", 60, 2, 12, verbose = 0)
jm_sarima <- janelamovel(AirPassengers, "sarima", 60, 2, 12, verbose = 1, refit.cada = 12)
jm_sarima <- janelamovel(AirPassengers, "sarima", 60, 2, 12, verbose = 2, refit.cada = 3)
}

}
