% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod-reg_lin.r
\name{modelos_reg_lin}
\alias{modelos_reg_lin}
\alias{reg_lin}
\alias{predict.reg_lin}
\title{Modelos \code{reg_lin}}
\usage{
reg_lin(serie, regdata, formula, ...)

\method{predict}{reg_lin}(object, newdata, n.ahead, ...)
}
\arguments{
\item{serie}{série para ajustar}

\item{regdata}{\code{data.frame}-like contendo variáveis explicativas}

\item{formula}{opcional, fórmula da regressão. Se for omitido, todas as variaveis em
\code{regdata} serão utilizadas}

\item{...}{demais argumentos passados a \link[stat]{\code{predict.lm}}}

\item{object}{objeto com classes \code{c("reg_lin", "modprev")} contendo modelo}

\item{newdata}{\code{data.frame}-like contendo variéveis explicativas fora da amostra}

\item{n.ahead}{número de passos à frente para previsão. Este argumento não é necessario, caso não
seja informado a previsão sera feita tantos passos à frente quanto amostras em \code{newdata}}
}
\value{
Objeto da classe \code{modprev} e subclasse \code{reg_lin}, uma lista de dois
elementos: \code{modelo} e \code{serie} contendo o modelo estimado e a série passada

\code{predict} serie temporal multivariada contendo valor esperado e desvio padrao de
previsão \code{n.ahead} passos à frente;
}
\description{
Estimação e métodos de modelos da classe \code{reg_lin}
}
