% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod-reg_lin.r
\name{modelos_reg_lin}
\alias{modelos_reg_lin}
\alias{reg_lin}
\alias{predict.reg_lin}
\alias{update.reg_lin}
\title{Modelos \code{reg_lin}}
\usage{
reg_lin(serie, regdata, formula, pesos = rep(1, length(serie)), ...)

\method{predict}{reg_lin}(object, newdata, n.ahead, ...)

\method{update}{reg_lin}(object, newseries, newregdata, refit = FALSE, ...)
}
\arguments{
\item{serie}{série para ajustar}

\item{regdata}{\code{data.frame}-like contendo variáveis explicativas}

\item{formula}{opcional, fórmula da regressão. Se for omitido, todas as variaveis em
\code{regdata} serão utilizadas}

\item{pesos}{opcional, vetor de pesos para cada observacao no ajuste. Sera normalizado
internamente}

\item{...}{nao possui uso, existe apenas para consistencia com a generica}

\item{object}{objeto com classes \code{c("reg_lin", "modprev")} contendo modelo}

\item{newdata}{\code{data.frame}-like contendo variéveis explicativas fora da amostra}

\item{n.ahead}{número de passos à frente para previsão. Este argumento não é necessario, caso não
seja informado a previsão sera feita tantos passos à frente quanto amostras em \code{newdata}}

\item{newseries}{nova série com a qual atualizar o modelo}

\item{newregdata}{\code{data.frame}-like contendo variáveis explicativas na nova amostra}

\item{refit}{booleano indicando se o modelo deve ou nao ser reajustado}
}
\value{
Objeto da classe \code{modprev} e subclasse \code{reg_lin}, uma lista de dois
elementos: \code{modelo} e \code{serie} contendo o modelo estimado e a série passada

\code{predict} serie temporal multivariada contendo valor esperado e desvio padrao de
previsão \code{n.ahead} passos à frente;

\code{update} retorna modelo com novos dados e, caso \code{refit == TRUE}, reajustado.
Contrário à função de estimação, \code{update} já retorna o objeto da classe \code{modprev};
}
\description{
Estimação e métodos de modelos da classe \code{reg_lin}

Para estimação deste tipo de modelo, além de ser fornecido o arumento \code{serie} também deve
ser necessariamente informado \code{regdata}, um \code{data.frame}-like contendo as variáveis
explicativas necessárias.

A previsão destes modelos é feita com um argumento opcional já passado por padrão que não pode
ser modificado. Este é \code{se.fit = TRUE}, determinando que seja retornado o desvio padrão
associado a previsão realizada. Considerando estes fatores, \code{...} não pode conter
\code{n.ahead} ou estes dois outros, ou então ocorrerá erro.
}
\details{
Modelos de regressao linear estaticas comuns, estimados atraves da funcao \code{\link[stats]{lm}}

Opcionalmente pode ser informado o argumento \code{formula} contendo a especificação da regressão
linear, no formato padrão do R (veja \code{\link{formula}}) porém sem o LHS. Caso \code{formula}
seja omitido, todas as variáveis em \code{regdata} serão utilizadas aditivamemte, i.e. se existem
as colunas \code{c("V1", "V2", "V3")}, formula sera \code{~ V1 + V2 + V3}.

Na execucao em janela movel, o vetor de pesos deve ser passado com o tamanho igual ao da janela.
O mesmo vetor sera aplicado em todas as janelas avaliadas.


}
