% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod-sarima.r
\name{modelos_sarima}
\alias{modelos_sarima}
\alias{sarima}
\alias{predict.sarima}
\alias{update.sarima}
\title{Modelos \code{sarima}}
\usage{
sarima(serie, regdata, formula)

\method{predict}{sarima}(object, n.ahead, ...)

\method{update}{sarima}(object, newseries, refit = FALSE, ...)
}
\arguments{
\item{serie}{serie para ajustar}

\item{regdata}{opcional, \code{data.frame}-like contendo variáveis explicativas}

\item{formula}{opcional, fórmula da regressão. Se for omitido, todas as variaveis em
\code{regdata} serão utilizadas. So tem uso se \code{regdata} for passado}

\item{object}{objeto com classes \code{c("sarima", "modprev")} contendo modelo}

\item{n.ahead}{número de passos à frente para previsão}

\item{...}{existe apenas para consistência com a genérica}

\item{newseries}{nova série com a qual atualizar o modelo}

\item{refit}{booleano indicando se o modelo deve ou não ser reajustado}
}
\value{
Objeto da classe \code{modprev} e subclasse \code{sarima}, uma lista de dois elementos:
\code{modelo} e \code{serie} contendo o modelo estimado e a série passada

\code{predict} retorna uma série temporal multivariada contendo valor esperado e desvio
padrão da previsão \code{n.ahead} passos à frente;

\code{update} retorna modelo com novos dados e, caso \code{refit == TRUE}, reajustado.
Contrário à função de estimação, \code{update} já retorna o objeto da classe \code{modprev};
}
\description{
Estimação e métodos de modelos da classe \code{sarima}
}
\details{
Esta especificação identifica e ajusta um modelo SARIMA(p, d, q)(P, D, Q)^s através da função
\code{\link[forecast]{auto.arima}}, que realiza uma busca no espaço de hiperparâmetros de ordem
do modelo por aquele conjunto com menor BIC.
}
