% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod-ss_ar1_saz.r
\name{modelos_ss_ar1_saz}
\alias{modelos_ss_ar1_saz}
\alias{ss_ar1_saz}
\alias{predict.ss_ar1_saz}
\alias{update.ss_ar1_saz}
\title{Modelos \code{ss_ar1_saz}}
\usage{
ss_ar1_saz(serie, ...)

\method{predict}{ss_ar1_saz}(object, n.ahead, ...)

\method{update}{ss_ar1_saz}(object, newseries, refit = FALSE, ...)
}
\arguments{
\item{serie}{serie para ajustar}

\item{object}{objeto com classes \code{c("ss_ar1_saz", "mod_eol")} contendo modelo}

\item{n.ahead}{numero de passos a frente para previsao}

\item{newseries}{nova serie com a qual atualizar o modelo}

\item{refit}{booleano indicando se o modelo deve ou nao ser reajustado}
}
\value{
\code{ss_ar1_saz} retorna modelo espaco de estados AR(1) + SAZ estimado

\code{predict} serie temporal multivariada contendo valor esperado e desvio padrao de
previsao;

\code{update} retorna modelo com novos dados e, caso \code{refit == TRUE}, reajustado;
}
\description{
Estimacao e metodos de modelos da classe \code{ss_ar1_saz}
}
\details{
\bold{Update}:

A atualizacao de modelos \code{ss_ar1_saz} sempre vai checar se o modelo passado foi estimado
corretamente. Como modelos em espaco de estados dependem bastante de inicializacao, as vezes nao
da pra estimar direito. Nesses casos ele tenta reestimar o modelo independentemente de
\code{refit}
}
