% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod-ss_reg_din.r
\name{modelos_ss_reg_din}
\alias{modelos_ss_reg_din}
\alias{ss_reg_din}
\alias{predict.ss_reg_din}
\alias{update.ss_reg_din}
\title{Modelos \code{ss_reg_din}}
\usage{
ss_reg_din(serie, regdata, formula)

\method{predict}{ss_reg_din}(object, newdata, n.ahead, ...)

\method{update}{ss_reg_din}(object, newseries, newregdata, refit = FALSE, ...)
}
\arguments{
\item{serie}{serie para ajustar}

\item{regdata}{vetor, matriz ou data.frame contendo variaveis explicativas}

\item{formula}{opcional, formula da regressao. Se for omitido, todas as variaveis em
\code{regdata} serao utilizadas}

\item{object}{objeto com classes \code{c("ss_reg_din", "mod_eol")} contendo modelo}

\item{newdata}{vetor, matriz ou data.frame contendo variaveis explicativas fora da amostra}

\item{n.ahead}{numero de passos a frente para previsao. Este argumento nao e necessario, caso nao
seja informado a previsao sera feita tantos passos a frente quanto amostras em \code{newdata}}

\item{...}{demais argumentos passados a \code{\link[KFAS]{predict.SSModel}}}

\item{newseries}{nova serie com a qual atualizar o modelo}

\item{newregdata}{vetor, matriz ou data.frame contendo variaveis explicativas na nova amostra}

\item{refit}{booleano indicando se o modelo deve ou nao ser reajustado}
}
\value{
\code{ss_reg_din} retorna modelo de regressao dinamica simples

\code{predict} serie temporal multivariada contendo valor esperado e desvio padrao de
previsao;

\code{update} retorna modelo com novos dados e, caso \code{refit == TRUE}, reajustado;
}
\description{
Estimacao e metodos de modelos da classe \code{ss_reg_din}
}
\details{
\bold{Update}:

A atualizacao de modelos \code{ss_reg_din} sempre vai checar se o modelo passado foi estimado
corretamente. Como modelos em espaco de estados dependem bastante de inicializacao, as vezes nao
da pra estimar direito. Nesses casos ele tenta reestimar o modelo independentemente de
\code{refit}
}
