% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizacao.r
\name{plot.mod_eol}
\alias{plot.mod_eol}
\title{Plot de mod_eol}
\usage{
\method{plot}{mod_eol}(fit, ..., legend = TRUE)
}
\arguments{
\item{fit}{\link{objeto mod_eol} objeto para plotar}

\item{...}{demais parametros repassados para \code{plot}}
}
\description{
Metodo para plotar ajustes mod_eol
}
