% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classe-modprevU.r
\name{predict.modprevU}
\alias{predict.modprevU}
\title{Previsão De Modelos \code{modprevU}}
\usage{
\method{predict}{modprevU}(object, n.ahead, ...)
}
\arguments{
\item{object}{modelo ajustado através de \code{\link{estimamodeloU}}}

\item{n.ahead}{número de passos à frente para prever}

\item{...}{existe apenas para consistência com a genérica}
}
\value{
série temporal multivariada contendo a previsão e o desvio padrão associado para os
passos de tempo \code{1:n.ahead}
}
\description{
Wrapper para previsão de modelos unicos
}
\seealso{
Other Metodos modprevU: 
\code{\link{estimamodelo_U}()},
\code{\link{update.modprevU}()}
}
\concept{Metodos modprevU}
