% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classe-mod_eol.r
\name{update.mod_eol}
\alias{update.mod_eol}
\title{Atualizacao De Modelos \code{mod_eol}}
\usage{
\method{update}{mod_eol}(object, newseries, refit = FALSE, ...)
}
\arguments{
\item{object}{modelo ajustado atraves de \code{estimamodelo}}

\item{newseries}{nova serie para associar ao modelo}

\item{refit}{booleano indicando se o modelo deve ser reajustado}

\item{...}{existe apenas para consistência com a genérica}
}
\value{
modelo com novos dados, possivelmente reajustado
}
\description{
Wrapper para atualizar e possivelmete reajustar modelos \code{mod_eol}
}
\details{
O padrão desta função é simplesmente substituir \code{newseries} no modelo ajustado
\code{object}, isto é, mantendo todos os hiperparâmetros estimados originalmente. Através do
argumento \code{refit} é possível realizar o reajuste do modelo para a nova série.

\bold{Modelos com variáveis explicativas:}

No caso de modelos com variáveis explicativas, deve ser fornecido um argumento \code{newregdata}
na forma de um \code{data.frame}-like contendo as variáveis explicativas necessárias associadas
às novas observações em \code{newseries}.
}
\examples{

serie1 <- window(datregdin[[1]], 1, 300)
serie2 <- window(datregdin[[1]], 501, 800)

mod_orig  <- estimamodelo(serie1, tipo = "sarima")
mod_upd   <- update(mod_orig, serie2, refit = FALSE)
mod_refit <- update(mod_orig, serie2, refit = TRUE)

\dontrun{
layout(matrix(1:3))
plot(mod_orig)
plot(mod_upd)
plot(mod_refit)
}

# Com variaveis explicativas ---------------------------

serie <- window(datregdin[[1]], 1, 100)
varex <- datregdin[1:100, 2, drop = FALSE]
mod_orig <- estimamodelo(serie, "ss_reg_din", regdata = varex)

newserie <- window(datregdin[[1]], 101, 200)
newvarex <- datregdin[101:200, 2, drop = FALSE]

mod_upd   <- update(mod_orig, newserie, newregdata = newvarex, refit = FALSE)
mod_refit <- update(mod_orig, newserie, newregdata = newvarex, refit = TRUE)

\dontrun{
layout(matrix(1:3))
plot(mod_orig)
plot(mod_upd)
plot(mod_refit)
}

}
\seealso{
Other Metodos mod_eol: 
\code{\link{estimamodelo}()},
\code{\link{predict.mod_eol}()}
}
\concept{Metodos mod_eol}
