% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/janelamovel.r
\name{whichreturn}
\alias{whichreturn}
\title{Tipo De \code{return} Em \code{janelamovel}}
\usage{
whichreturn(full.output)
}
\arguments{
\item{full.output}{booleano indicando o tipo de retorno}
}
\value{
funcao de retorno para usar no lapply de \code{janelamovel}
}
\description{
Seleciona se a funcao de \code{return} no lapply devolve tudo ou so a previsao
}
